Completed refactoring of MultiChatSpigot
